+++ integreat_cms/cms/models/pois/poi.py
@@ -99,6 +99,12 @@ class POI(AbstractContentModel):
         verbose_name=_("Stops in the vicinity"),
         help_text=_("List here the stops for public transport in the vicinity. Please separate them by comma")
     )
+    spoken_languages = models.CharField(
+        max_length=50,
+        blank=True,
+        verbose_name=_("Spoken languages"),
+        help_text=_("List here all languages, that are spoken at the location. Please separate them by comma")
+    )
 
     @property
     def fallback_translations_enabled(self):